Repackage data-related files
It was done to separate the layers by implementing a repository in between.
Having a single Model object when it consists of two separate objects (User and Board) is a bad practice, so I extracted User.
