+++ app/src/main/kotlin/com/github/oryanmat/trellowidget/data/model/User.kt
@@ -1,0 +1,8 @@
+package com.github.oryanmat.trellowidget.data.model
+
+data class User(
+        val id: String = "",
+        val fullName: String = "",
+        val username: String = "") {
+    override fun toString() = "$fullName (@$username)"
+}
