+++ app/src/main/kotlin/com/github/oryanmat/trellowidget/widget/CardRemoteViewFactory.kt
@@ -12,14 +12,14 @@ import androidx.annotation.ColorInt
 import androidx.annotation.DrawableRes
 import androidx.annotation.IdRes
 import com.github.oryanmat.trellowidget.R
-import com.github.oryanmat.trellowidget.model.BoardList
-import com.github.oryanmat.trellowidget.model.Card
-import com.github.oryanmat.trellowidget.model.Label
+import com.github.oryanmat.trellowidget.data.model.BoardList
+import com.github.oryanmat.trellowidget.data.model.Card
+import com.github.oryanmat.trellowidget.data.model.Label
 import com.github.oryanmat.trellowidget.util.DateTimeUtil
 import com.github.oryanmat.trellowidget.util.RemoteViewsUtil.setImage
 import com.github.oryanmat.trellowidget.util.RemoteViewsUtil.setImageViewColor
 import com.github.oryanmat.trellowidget.util.RemoteViewsUtil.setTextView
-import com.github.oryanmat.trellowidget.util.TrelloAPIUtil
+import com.github.oryanmat.trellowidget.data.remote.TrelloApi
 import com.github.oryanmat.trellowidget.util.color.colors
 import com.github.oryanmat.trellowidget.util.color.dim
 import com.github.oryanmat.trellowidget.util.getCardForegroundColor
@@ -33,7 +33,7 @@ class CardRemoteViewFactory(private val context: Context,
 
     override fun onDataSetChanged() {
         var list = context.getList(appWidgetId)
-        list = TrelloAPIUtil.instance.getCards(list)
+        list = TrelloApi.instance.getCards(list)
         color = context.getCardForegroundColor()
 
         if (BoardList.ERROR != list.id) {
